Pertemuan 4 - DetailScreen

+++ film/lib/models/movie.dart
@@ -7,15 +7,14 @@ class Movie {
   final String releaseDate;
   final double voteAverage;
 
-  Movie({
-    required this.id,
-    required this.title,
-    required this.overview,
-    required this.posterPath,
-    required this.backdropPath,
-    required this.releaseDate,
-    required this.voteAverage
-  });
+  Movie(
+      {required this.id,
+      required this.title,
+      required this.overview,
+      required this.posterPath,
+      required this.backdropPath,
+      required this.releaseDate,
+      required this.voteAverage});
 
   factory Movie.fromJson(Map<String, dynamic> json) {
     return Movie(
@@ -25,7 +24,7 @@ class Movie {
       posterPath: json['poster_path'],
       backdropPath: json['backdrop_path'],
       releaseDate: json['release_date'],
-      voteAverage: json['vote_average'],
+      voteAverage: json['vote_average'].toDouble(),
     );
   }
 }
